feat: Implement basic liveness and readiness probes
This commit implements basic liveness and readiness probes for the application.

The liveness probe now always returns true.
The readiness probe returns the value of the `isReady` flag.
The `setReadiness` method now sets the `isShuttingDown` flag to true when `isReady` is set to false.
The try-catch block in `checkLiveness` has been removed as it was unnecessary.