feat: Add Swagger specification file
This commit introduces a Swagger specification file (`swaggerSpec.js`) to define the API documentation for the Event Spawning service.

The file utilizes `swagger-jsdoc` to generate the Swagger specification based on JSDoc comments within the `./src/modules/**/*.js` files.

The specification includes basic information about the API, such as title, version, and description, as well as a server configuration pointing to the local development environment.