feat: Remove default resources and ulimits, apply them directly to services
This commit removes the `x-default-resources` and `x-default-ulimits` YAML anchors from the `docker-compose.yml` file and applies the resource limits and ulimits directly to each service definition.

This change improves readability and maintainability by making the resource configuration explicit for each service. It avoids the potential for unintended sharing or modification of default values across services. Also, add nats as dependency for reporter.